API Functionality 4
Added a test shipment that successfully passes through the API and creates a "Shipment" with API-generates rates. The API key and Zip codes are missing, and have to be added manually. Still need to rename a lot of variables (such as street1 to streetOne) and organize the code to be more readable.